Conflicts:
	renrenapi/src/main/java/com/renrentui/renrenapi/dao/impl/ClienterDao.java
	renrenapi/src/main/resources/conf/mybatis/ClienterMapper.xml

+++ renrenapi/src/main/java/com/renrentui/renrenapi/service/impl/ClienterService.java
@@ -73,7 +73,7 @@ public class ClienterService implements IClienterService{
 	 * WangChao
 	 */
 	@Override
-	public int signup(SignUpReq req) {
+	public long signup(SignUpReq req) {
 		return clienterDao.signup(req); 
 	} 
 	/**
