+++ renrenapi/src/main/java/com/renrentui/renrenapi/service/inter/IClienterService.java
@@ -49,7 +49,7 @@ public interface IClienterService {
 	 */
 	boolean modifyPwdUserc(ModifyPwdReq req);
 
-	int signup(SignUpReq req);
+	long signup(SignUpReq req);
 	/**
 	* @Des 查询C端用户信息  
 	* @Author WangXuDan
