+++ renrenentity/src/main/java/com/renrentui/renrenentity/req/SignUpReq.java
@@ -1,19 +1,18 @@
 package com.renrentui.renrenentity.req;
 
 public class SignUpReq {
-	private Integer id;
+	private long id;
 	private String phoneNo;//手机号
 	private String passWord;//密码
 	private String verifyCode;//验证码
 	private String name;  //用户名
-	
-	public Integer getId() {
-        return id;
-    } 
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
+	 
+	public long getId() {
+		return id;
+	}
+	public void setId(long id) {
+		this.id = id;
+	}
 	public String getPhoneNo() {
 		return phoneNo;
 	}
